requirements.txt
sklearn
torchvision
torchaudio
sox
plotly
scikit-image
